bugfix/efmls-doesnt-load-by-mason - Separate each setup in their own file
