+++ config/nvim/lua/plugins/nvim-lspconfig/setups/mason.lua
@@ -1,0 +1,11 @@
+local mason                = require 'mason'
+local mason_lspconfig      = require 'mason-lspconfig'
+local mason_tool_installer = require("mason-tool-installer")
+
+local function setup_mason()
+  mason.setup()
+  mason_tool_installer.setup { automatic_installation = true }
+  mason_lspconfig.setup { automatic_installation = true }
+end
+
+return setup_mason
