+++ config/nvim/lua/plugins/nvim-lspconfig/solargraph_config.lua
@@ -1,0 +1,12 @@
+local default_config = require('plugins.nvim-lspconfig.default_config')
+
+return {
+  capabilities = default_config.capabilities,
+  settings = {
+    solargraph = {
+      diagnostics = true,
+      completion = true,
+      formatting = true
+    }
+  }
+}
